requirements.txt
streamlit
PyMuPDF
mysql-connector-python
transformers
torch
pandas
python-dotenv